+++ packages/expo-localization/CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### 💡 Others
 
+- [Web] Add `getTextInfo()` fallback for deprecated `textInfo` in `getLocales()` ([#39255](https://github.com/expo/expo/pull/39255) by [@desii101](https://github.com/desii101))
+
 ## 17.0.7 — 2025-09-11
 
 _This version does not introduce any user-facing changes._